Add checks for zstd tarball extraction from ONIE or Aboot
Currently, there are two methods of having a reduced image size:
BUILD_REDUCE_IMAGE_SIZE build configuration (resulting in the use of a
zstd-compressed docker tarball for the installation file), and
docker_inram (resulting in the docker tarball being extracted into RAM
at bootup time instead to disk) (this particular option seems to be
better supported on Aboot and not ONIE). These two are logically
separate; one is a build-time configuration for an image, and the other
is a runtime platform-specific option.

This means that it's theoretically possible to have an image be built
with BUILD_REDUCE_IMAGE_SIZE enabled and then be loaded onto a device
that doesn't use docker_inram. If the image is being installed from
SONiC, then that's _probably_ file, as long as the SONiC installation
has zstd installed. If the image is being installed from ONIE or Aboot,
then both of those environments currently don't support zstd-compressed
tarballs, so image installation fails.

Because of that, fail explicitly saying that this image is not supported
in that environment.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>