Fix the issue of 'Peer-group members must be all internal or all external'
The issue is seen on t0-d18u8s4. The peer switches (internal) and t1 switches
(external) were added to the same BGP peer group, which is not allowed.

The error is from bgpcfgd:
Peer-group members must be all internal or all external

Skip adding the internal neighbors (same ASN) to peer group.