Update Dockerfile.j2
The latest version of setuptools-scm i.e. 8.2.0 does not support bdist_wheel. Hence there is a failure in target docker-syncd-rpc.gz due to ptf-py3 build failure. On further investigation found that bookworm installs 8.0.4 version of this package. Hence added this to the current Dockerfile.j2 of sonic-slave-bullseye. This will make it consistent with the earlier build process for ptf-py3 which used 8.1.0 version of setuptools-scm and also the sonic-slave-bookworm