Add link requirement per sku before enabling traffic on a fap (#21865)
With the current fap initialization sequence, multiple faps in a modular system could come up with stand alone mode. This is not a recommended configuration. It might mess up the global clock and might ends up with traffic issue.
#21866

This change adds a required minimum number of links have to be ready before enabling traffic on a fap, which based on the guidance should be able to guarantee the faps in a modular system can form a proper topology.