loadData: do not mutate user-provided data
The user-provided configdb should not be mutated as the user
may expect it to be unmodified.  It is fairly unlikely to need
to modify the input data, and it doesn't make sense to force
the caller to deep copy the data before passing it in just in
case it gets modified.