Adding 'custom_feature_start_tx_threshold=20' on Arista SKUs
It's theorized that https://github.com/aristanetworks/sonic/issues/120
could be due to a FIFO underflow which is documented here
https://brcmsemiconductor-csm.wolkenservicedesk.com/wolken-support/article?articleId=19817
In EOS we workaround this FIFO underflow issue by setting the
start_tx_threshold=20 for 100G and 400G ports. Setting the same value
via the soc properties in this change to match EOS values.