Merged PR 11968083: Add the missing start.sh in the docker image for dummy container.
Add the missing start.sh in the docker image for dummy container.

Without start.sh the latest dummy container won't be able to be started
by k8s, which caused rollback in prod.

----
#### AI description  (iteration 1)
#### PR Classification
Bug fix

#### PR Summary
This pull request adds the missing `start.sh` script to the Docker image for the dummy container.
- `dockers/docker-dummyk8s/Dockerfile.j2`: Added `start.sh` to the list of files copied into the Docker image.
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->