implement configdb<>xpath conversion
Import modified xpath to configdb path and configdb path to xpath
helpers from sonic-utilites/generic_config_updater as they don't
belong there.  Modify the code to be more readable, add the ability
to extract schema paths instead of data xpaths. Don't require
a configdb to be passed in unless we need to know a list index.

This also caches configdb->xpath conversion which is used in
recursive searches as it may be an expensive operation.