[bgpcfgd]: Delete "listen range” before attempting to delete a “peer group”.
Assume we have a BGP peer group called `BGPSLBPassive`.
This peer group is attached to a listen range `192.168.0.0/21`.

CONFIG_DB configuration:

> $ sonic-db-cli CONFIG_DB hgetall "BGP_PEER_RANGE|BGPSLBPassive"
> {'ip_range@': '192.168.0.0/21', 'name': 'BGPSLBPassive', 'src_address': '10.1.0.32'}

FRR configuration:

> ...
> router bgp 65100
>  neighbor BGPSLBPassive peer-group
>  bgp listen range 192.168.0.0/21 peer-group BGPSLBPassive
> ...

Now, we delete the peer group `BGPSLBPassive` from CONFIG_DB:

> $ sonic-db-cli CONFIG_DB del BGP_PEER_RANGE|BGPSLBPassive

In response to this, bgpcfgd tries to run the following vtysh command
to delete the peer group from the FRR configuration:

> vtysh -c "no neighbor BGPSLBPassive peer-group"

But the above vtysh command returns an error:

> %%Peer-group BGPSLBPassive is attached to 1 listen-range(s), delete them first

The problem is due to a change in recent versions of FRR.

Starting with FRR 10.1, if a peer group is attached to a "listen range",
the range must be removed before the peer group can be deleted.

In order to fix this issue, this commit does a change in bgpcfgd.

When bgpcfgd has to delete a peer group, it first runs the command
"no bgp listen range ..." to remove the "listen range" associated with
the peer group, and only then proceed with deleting the peer group.

Signed-off-by: Carmine Scarpitta <cscarpit@cisco.com>