Update version files to allow pygobject 3.42.2 to be used
Even after installing python3-gi, pygobject version 3.50.0 continues to
be installed from the pypi repo when installing sonic-host-services.
This is because the version control files specify that version 3.50.0 of
this package should be used, and pip complies with that, installing it
from the pypi repo.

Modify the version control files to let 3.42.2 be used.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>