Merged PR 12372008: Add MAIA200 resource type in BGP template
Add MAIA200 resource type in BGP template

----
#### AI description  (iteration 1)
#### PR Classification
This pull request adds a new feature by incorporating support for the AI-MS200 resource type in the BGP template configuration.

#### PR Summary
The pull request updates the BGP template to set the `is_fairwater` flag when the `resource_type` is defined as `AI-MS200` alongside the existing check for `FairWater`.
- `/dockers/docker-fpm-frr/frr/bgpd/templates/msft.general/router.j2`: Revised the conditional logic to include the new `AI-MS200` resource type.
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->