Merged PR 12407212: Dummy docker improvements
Add dummy docker into build target so that dummy docker is built and pushed to ACR (but service is not running on SONiC)

----
#### AI description  (iteration 1)
#### PR Classification
This PR introduces a new feature by always including the dummy docker image in the build target.

#### PR Summary
The changes simplify the build rule for the dummy docker by removing the conditional statements, ensuring its inclusion regardless of configuration flags.
- `rules/docker-dummyk8s.mk`: Removed the `ifeq ($(INCLUDE_DUMMYK8S),y)` block and corresponding `endif`, and consistently adds the dummy docker image to the build.
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->