Set admin_status down to ports not part of minigraph
Why I did it
When minigraph is deployed, config_db has "admin_status": "up"
for ports that are required for the specific topo in minigraph.xml.
Ports not included in the topology are left without an
"admin_status" field.

This leads to a problem where buffermgrd waits indefinitely for
the "admin_status" field to be set before applying the pg_lossless profile.

As a result, certain tests in sonic-mgmt errors out during teardown,
particularly those involving interface admin status changes and rollbacks.

How I did it
During minigraph deploy set "admin_status" to "down" for ports
that are not part of topology in minigraph.xml.

Which release branch to backport (provide reason below if selected)
* [ ]  202006
* [ ]  202012
* [ ]  202106
* [ ]  202111
* [ ]  202205
* [ ]  202211
* [ ]  202305
* [ ] 202311
* [ ] 202405
* [X] 202411

How to verify it:
Run below testcase on T0 topology:
generic_config_updater/test_eth_interface.py
::test_eth_interface_admin_change[down]

Signed-off-by: Ravi Minnikanti <rminnikanti@marvell.com>