Tell chrony that the RTC is always in UTC
Tell chrony that the RTC on devices is always stored in UTC and not the
local time zone. For devices that are running on UTC, this won't change
anything; for devices that are not running on UTC, this will correct
some time issues where the kernel interprets the time in the RTC as
being in UTC, but Chrony interprets that time as being in the local time
zone.

Fixes #22437

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>