YANG: PREFIX list implementation does not take sequence_id
The sequence id is used as a key in the yang model and is therefor
mandatory, however the implementation does not expect a sequence
id at all and just uses the order in which the prefixes appear.

If passing value as YANG requires, the jinja2 renderer and frrcfgd
both fail.

This updates the YANG model to match the implementation, including
test cases.

Signed-off-by: Brad House <bhouse@nexthop.ai>