[FRR]: Fix a BGP check that verifies if the nexthop is invalid
Currently, in `bgp_update_martian_nexthop` we do a check to evaluate whether a nexthop is valid or not. But then the result of this check is not saved and therefore it gets lost.

```
		if (!ipv4_unicast_valid(&attr->nexthop) ||
			      bgp_nexthop_self(bgp, afi, type, stype, attr, dest));
```

This commit fixes the issue by saving the result of the check in the `nh_invalid` variable that already exists.