libyang-python: fix test crash seen in Trixie
There is an internal Python CFFI pointer that gets invalidated in a
test case due to the internal object being mutated.  This test case
appears to be malformed and makes incorrect assumptions.

Nothing here in particular appears to be actually related to Trixie
so it isn't clear why it shows up. The feature being tested is also
in a code path that is not used at all by SONiC.

https://github.com/CESNET/libyang-python/pull/141

Signed-off-by: Brad House <brad@brad-house.com>