Add gbsyncd container for milleniob
#### Why I did it

This change is to support platforms that use the MillenioB (BCM81356) external PHY chips.

##### Work item tracking
- Microsoft ADO **(number only)**:

#### How I did it

The change introduces a new docker-gbsyncd-milleniob container, similar to the container for Broncos added in https://github.com/sonic-net/sonic-buildimage/pull/11154
The gbsyncd.ini file is added for determining which gbsyncd container a platform needs.
By default, the container will not be included in the built image unless the (LIBSAI_MILLENIOB)_URL is defined (pointing to a deb package containing the milleniob).

#### How to verify it

#### Which release branch to backport (provide reason below if selected)

- [ ] 201811
- [ ] 201911
- [ ] 202006
- [ ] 202012
- [ ] 202106
- [ ] 202111
- [ ] 202205
- [ ] 202211
- [ ] 202305
- [X] 202505

The platform that uses this phy is to be put in 202505.

#### Tested branch (Please provide the tested image version)

#### Description for the changelog

#### Link to config_db schema for YANG module changes

#### A picture of a cute animal (not mandatory but encouraged)