Loosen help text check, as it might change depending on versions
Loosen the help text check in `test_cfghelp.py` in `sonic-yang-mgmt`.
The exact text might change from one Python version to another, and help
text itself is more for use by a human rather than a machine. It's
better to check that the expected elements of the help text (something
about the options that are expected and the descriptions) are there
rather than the exact formatting.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>
Signed-off-by: xiaweijiang <xiaweijiang@microsoft.com>