Increase the installer disk size by 8MB for KVM
Following #22199, the debug version of `sonic-vs.bin` file is at 4093MB
(to be precise, one particular build was at 4291790831 bytes). This, in
theory, should've fit fine in the 4096MB FAT32 installer disk that is
created for building the sonic-vs.img QCOW2 image. However, it's short
by 5MB. My guess for this reason is because FAT32 itself needs some
space for metadata, and so we don't get the full 4096MB. This results in
a failure to build the debug KVM image.

As a fix, increase the installer disk size that we create by 8MB, which
should be enough for any debug `sonic-vs.bin` file that we create for
this filesystem. I say "for this filesystem" because FAT32 allows files
up to 4GB (minus one byte). This means that if we need to create a debug
image larger than this, then the installer disk will need to use a
different filesystem, probably ext2 or ext4. Considering the ONIE recovery
environment is a Linux environment, this will probably be perfectly
fine if we do need to change to one of those two filesystems, as long as
nothing has FAT32 hardcoded anywhere.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>